+++ app/models/activity.rb
@@ -192,6 +192,7 @@ class Activity < ApplicationRecord
   has_many :historical_events, dependent: :destroy
 
   has_one :commitment, dependent: :destroy
+  validates_associated :commitment
 
   has_many :reports,
     ->(activity) { unscope(:where).for_activity(activity).in_historical_order }