+++ spec/features/users_can_upload_activities_spec.rb
@@ -93,6 +93,13 @@ RSpec.feature "users can upload activities" do
       expect(page).to have_xpath("td[3]", text: "")
       expect(page).to have_xpath("td[4]", text: "Newton Fund Pillar can't be blank")
     end
+
+    within "//tbody/tr[4]" do
+      expect(page).to have_xpath("td[1]", text: "Original commitment figure")
+      expect(page).to have_xpath("td[2]", text: "")
+      expect(page).to have_xpath("td[3]", text: "-100")
+      expect(page).to have_xpath("td[4]", text: "Original commitment figure is invalid")
+    end
   end
 
   context "uploading a set of activities the user doesn't have permission to modify" do
@@ -339,6 +346,7 @@ RSpec.feature "users can upload activities" do
         "ISPF non-ODA partner countries",
         "Implementing organisation names",
         "Tags",
+        "Original commitment figure",
         "Comments"
       ])
     end
@@ -377,6 +385,9 @@ RSpec.feature "users can upload activities" do
       # Then I should see the tags
       expect(page).to have_content("Ayrton Fund")
       expect(page).to have_content("ICF Funded")
+
+      # And I should see the original commitment figure
+      expect(page).to have_content("Original commitment figure £1,000.00")
     end
 
     scenario "linking an activity to a non-ODA activity via the bulk upload" do
@@ -437,6 +448,7 @@ RSpec.feature "users can upload activities" do
         "ISPF non-ODA partner countries",
         "Implementing organisation names",
         "Tags",
+        "Original commitment figure",
         "Comments"
       ])
     end
@@ -475,6 +487,9 @@ RSpec.feature "users can upload activities" do
       # Then I should see the tags
       expect(page).to have_content("ICF Funded")
       expect(page).to have_content("Tactical Fund")
+
+      # And I should see the original commitment figure
+      expect(page).to have_content("Original commitment figure £1,000.00")
     end
 
     scenario "linking an activity to a ODA activity via the bulk upload" do
@@ -544,6 +559,7 @@ RSpec.feature "users can upload activities" do
         "UK PO Named Contact",
         "NF Partner Country PO",
         "Implementing organisation names",
+        "Original commitment figure",
         "Comments"
       ])
     end